% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checklist.R
\name{checklist}
\alias{checklist}
\alias{checklist.taxlist}
\alias{checklist,taxlist-method}
\title{Taxonomic Check-List}
\usage{
checklist(object, ...)

\method{checklist}{taxlist}(
  object,
  exclude,
  indent = "&nbsp;",
  prefix = "",
  suffix = "",
  print_args = list(),
  rmd_args = list(`header-includes` = "- \\\\setlength{\\\\parskip}{0pt}"),
  render_args = list(),
  title = "Taxonomic Checklist",
  output = "pdf_document",
  output_file = tempfile(),
  ...
)
}
\arguments{
\item{object}{A \linkS4class{taxlist} object used as main input.}

\item{...}{Further arguments. In taxlist-method they will be passed to the
function \code{\link[=indented_list]{indented_list()}}.}

\item{exclude}{A character vector including the name of taxonomic ranks that
will be excluded of formatting in italics (excluded in the call to
\code{\link[=print_name]{print_name()}}).}

\item{indent}{A character value indicating the single indentation, which
will be multiplied by the taxonomic rank going from the top to the
bottom. Indentation can be either cancellec by \code{indent = ""} or can be
specified for determined taxonomic ranks by naming elements of the input
vector after the ranks.}

\item{prefix}{A character value or character vector named by taxonomic ranks,
which will be used as prefix.}

\item{suffix}{The same as \code{prefix} but it will be placed at the end of the
line.}

\item{print_args}{A list with elements named after parameters (arguents) of
the function \code{\link[=print_name]{print_name()}}.}

\item{rmd_args}{A list with elements named after parameters (arguments) of
the function \code{\link[=write_rmd]{write_rmd()}}.}

\item{render_args}{A list with elements named after parameters (arguments) of
the function \code{\link[=render_rmd]{render_rmd()}}.}

\item{title}{A character value with the title of the document.}

\item{output}{A character value indicating the format of ouput using
rmarkdown.}

\item{output_file}{Name of the rendered otuput file. It will be passed to
\code{\link[=write_rmd]{write_rmd()}}.}
}
\description{
Template for a taxonomic check-list generated from \linkS4class{taxlist} objects.
The most important functions called in the source code are \code{\link[=print_name]{print_name()}},
\code{\link[=indented_list]{indented_list()}}, \code{\link[=write_rmd]{write_rmd()}}, and \code{\link[=render_rmd]{render_rmd()}}.
}
\examples{
\dontrun{
library(taxlist)

## Subset of family Piperaceae
Piperaceae <- subset(Easplist, TaxonName == "Piperaceae",
  slot = "taxonNames", keep_children = TRUE, keep_parents = TRUE
)

## Simple checklist
checklist(Piperaceae, output_file = "piperaceae", exclude = "family")

## Families as sections in document
checklist(Piperaceae,
  output_file = "piperaceae2", exclude = "family",
  prefix = c(family = "# ")
)

## Families as sections and small captions for the whole taxonomic list
checklist(Easplist,
  output_file = "ea-splist", exclude = "family",
  prefix = c(family = "# \\\\sc{"), suffix = c(family = "}")
)
}

}
